Change types to fix overflow

+++ solutions/project-euler/002/002.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 
 int main() {
-    int16_t a = 1, b = 2, aux, sum = 2;
+    int32_t a = 1, b = 2, aux, sum = 2;
     while (b < 4000000) {
         aux = a + b;
         a = b;
